# backend/medi-snow/docs/swagger.yaml
basePath: /
definitions:
  main.PingResponse:
    properties:
      message:
        description: Response message
        example: pong
        type: string
    type: object
  types.Coords:
    properties:
      latitude:
        example: 40.7128
        type: number
      longitude:
        example: -74.006
        type: number
    type: object
  types.Elevation:
    properties:
      feet:
        example: 5280
        type: number
      meters:
        example: 1609.34
        type: number
    type: object
  types.ForecastPoint:
    properties:
      coordinates:
        $ref: '#/definitions/types.Coords'
      elevation:
        $ref: '#/definitions/types.Elevation'
      location:
        $ref: '#/definitions/types.LocationInfo'
    type: object
  types.LocationInfo:
    properties:
      country:
        example: United States
        type: string
      country_code:
        example: US
        type: string
      county:
        example: Denver County
        type: string
      name:
        example: Denver
        type: string
      state:
        example: Colorado
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Weather and avalanche forecasting API for ski areas
  title: Medi-Snow API
  version: 1.0.0
paths:
  /location/forecast-point:
    get:
      consumes:
      - application/json
      description: Retrieve comprehensive location data including coordinates, elevation,
        and location metadata for a given latitude and longitude
      parameters:
      - description: Latitude in decimal degrees
        example: 39.11539
        in: query
        maximum: 90
        minimum: -90
        name: latitude
        required: true
        type: number
      - description: Longitude in decimal degrees
        example: -107.6584
        in: query
        maximum: 180
        minimum: -180
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ForecastPoint'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get forecast point data
      tags:
      - location
  /ping:
    get:
      description: Check if the API is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.PingResponse'
      summary: Ping health check
      tags:
      - health
swagger: "2.0"